requirements.txt
alabaster==0.7.10
amqp==2.2.2
Babel==2.5.3
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
coverage==4.5.1
Django==2.0.4
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-debug-toolbar==1.9.1
django-haystack==2.8.1
django-multiselectfield==0.1.8
django-nose==1.4.5
django-redis==4.9.0
django-sphinxdoc==1.5.1
docutils==0.14
idna==2.6
imagesize==1.0.0
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
nose==1.3.7
packaging==17.1
pycrypto==2.6.1
pygame==1.9.3
Pygments==2.2.0
pyparsing==2.2.0
python-redis-lock==3.2.0
pytz==2018.3
redis==2.10.6
redis-cache==0.1.4
redis-structures==0.1.7
requests==2.18.4
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.2
sphinxcontrib-django==0.3.1
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.4
urllib3==1.22
vine==1.1.4
vital-tools==0.1.10
